Refactor commands, small optimizations
