+++ src-tauri/src/lifecycle/cork_vote.rs
@@ -1,13 +1,13 @@
-use eyre::{bail, Context, Result};
+use eyre::{Context, Result};
 use somm_proto::cork::{
     query_client::QueryClient as CorkQueryClient, QueryCorkResultRequest, QueryCorkResultResponse,
 };
-use tauri::{async_runtime::Sender, Manager};
+use tauri::Manager;
 use tonic::Response;
 
 use crate::{
     app::AppContext,
-    state::{self, RequestStatus},
+    state::{self},
 };
 
 /// Monitor the x/cork module for the result of the cork vote
