+++ src-tauri/src/lifecycle/relay.rs
@@ -1,18 +1,4 @@
-use std::str::FromStr;
-
-use alloy::{
-    contract::Event,
-    primitives::address,
-    providers::{Provider, ProviderBuilder},
-    rpc::types::Filter,
-    sol,
-};
-use alloy_primitives::FixedBytes;
-use eyre::Result;
-use tauri::async_runtime::Sender;
-use toml::from_str;
-
-use crate::{app::AppContext, state::RequestStatus};
+use alloy::sol;
 
 const GRAVITY_CONTRACT_ADDRESS: &str = "0x69592e6f9d21989a043646fe8225da2600e5a0f7";
